+++ synccommerce/RELEASES.md
@@ -571,3 +571,13 @@ Based on:
 - [typescript v0.32.1] synccommerce
 ### Releases
 - [NPM v0.32.1] https://www.npmjs.com/package/@codat/sync-for-commerce/v/0.32.1 - synccommerce
+
+## 2023-08-31 00:01:19
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce.yaml
+- Speakeasy CLI 1.75.0 (2.89.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.33.0] synccommerce
+### Releases
+- [NPM v0.33.0] https://www.npmjs.com/package/@codat/sync-for-commerce/v/0.33.0 - synccommerce